% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highexpressed.R
\name{highexpressed}
\alias{highexpressed}
\alias{high}
\alias{expressed}
\alias{genes}
\alias{libraries}
\title{highexpressed}
\usage{
highexpressed(obj, count_zeros = TRUE, num_max = 20)
}
\arguments{
\item{obj}{list object from the function ReadData}

\item{count_zeros}{boolean of if the number of zero for each library must be shown. Default: TRUE}

\item{num_max}{int of maximum number of genes reported on high expressed genes. To show all genes expressed use Inf}
}
\value{
list of the high expressed genes
}
\description{
Return the highest expressed genes from libraries or treatments.
}
\examples{
not run
dir <-system.file("extdata", package="mnmer")
dat <- ReadData (counts, expdata, design="~0+trats")
tgs <- highexpressed (dat)

}
\keyword{High}
\keyword{expressed}
\keyword{genes}
\keyword{high}
\keyword{variance}
