% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highvar.R
\name{highvar}
\alias{highvar}
\alias{high}
\alias{variance}
\alias{cluster}
\alias{PCA}
\alias{gene}
\alias{expression}
\title{highvar}
\usage{
highvar(obj, type = "PCA")
}
\arguments{
\item{obj}{list object from the function ReadData}

\item{types}{character of type of nalysis to be performed}
}
\value{
list of the treatments and its varince to compared
}
\description{
Analyse if a experiments has a high variance among replicates and treatments.
}
\examples{
not run
dir <-system.file("extdata", package="mnmer")
dat <- ReadData (counts, expdata, design="~trats")
tgs <- varanalysis (dat)

}
\keyword{PCA}
\keyword{association}
\keyword{cluster}
\keyword{high}
\keyword{variance}
