% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaanalysis.R
\name{pcaanalysis}
\alias{pcaanalysis}
\alias{principal}
\alias{component}
\alias{analysis}
\alias{gene}
\alias{expression}
\title{pcaanalysis}
\usage{
pcaanalysis(obj, plotcomps = TRUE, plotpcacor = TRUE)
}
\arguments{
\item{obj}{list object from the function ReadData}

\item{num}{int of number of standard deviation reported. Default: 20. To show all genes expressed use Inf}
}
\value{
list of the variane by genes among treatments
}
\description{
Return the variance analysis including near zero variance with treatments.
}
\examples{
not run
dir <-system.file("extdata", package="mnmer")
dat <- ReadData (counts, expdata, design="~trats")
tgs <- varanalysis (dat)
}
\keyword{High}
\keyword{expressed}
\keyword{genes}
\keyword{high}
\keyword{variance}
