% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadData.R
\name{ReadData}
\alias{ReadData}
\alias{count}
\alias{matrix}
\alias{read}
\alias{design}
\title{ReadData}
\usage{
ReadData(counts, col_data, design, norm = "scale", min_counts = 0)
}
\arguments{
\item{counts}{matrix type containing the gene as rows names and the libraries names as colunms}

\item{col_data}{data.frame containg the design table of the experiments}

\item{design}{formula of the factors and treatments}

\item{norm}{character of the normalization methods: scale (default), TMM, DESeq2}

\item{min_counts}{minimal row sum that is accpeted by further analisis. Default: 0}
}
\value{
list of objects related to
}
\description{
Read data from count matrix and experimental design file.
}
\examples{
not run
dir <-system.file("extdata", package="mnmer")
dat <- ReadData (counts, expdata, design="~trats")

}
\keyword{High}
\keyword{formulae}
\keyword{replicates}
\keyword{variance}
