% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varanalysis.R
\name{varanalysis}
\alias{varanalysis}
\alias{variance}
\alias{standard}
\alias{deviation}
\alias{gene}
\alias{expression}
\title{varanalysis}
\usage{
varanalysis(obj, num = 20)
}
\arguments{
\item{obj}{list object from the function ReadData}

\item{num}{int of number of standard deviation reported. Default: 20. To show all genes expressed use Inf}
}
\value{
list of the variane by genes among treatments
}
\description{
Return the variance analysis including near zero variance with treatments.
}
\examples{
not run
dir <-system.file("extdata", package="mnmer")
dat <- ReadData (counts, expdata, design="~trats")
tgs <- varanalysis (dat)

}
\keyword{High}
\keyword{expressed}
\keyword{genes}
\keyword{high}
\keyword{variance}
